+++ dart/lib/src/sentry.dart
@@ -94,6 +94,14 @@ class Sentry {
       options.addIntegrationByIndex(0, IsolateErrorIntegration());
     }
 
+    if (options.platformChecker.isDebugMode()) {
+      options.debug = true;
+      options.logger(
+        SentryLevel.debug,
+        'Debug mode is enabled: Application is running in a debug environment.',
+      );
+    }
+
     if (options.enableDartSymbolication) {
       options.addIntegration(LoadDartDebugImagesIntegration());
     }